Refactor selection tests
