+++ tests/test_amulet_core/test_selection/__init__.pyi
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from . import (
+    test_selection_box_,
+    test_selection_box_group_,
+    test_selection_cuboid_,
+    test_selection_ellipsoid_,
+    test_selection_shape_group_,
+)
+
+__all__: list[str] = [
+    "test_selection_box_",
+    "test_selection_box_group_",
+    "test_selection_cuboid_",
+    "test_selection_ellipsoid_",
+    "test_selection_shape_group_",
+]
