+++ tests/test_amulet_core/test_selection/test_selection_ellipsoid_.pyi
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+import collections.abc
+
+__all__: list[str] = ["get_tests"]
+
+def get_tests() -> list[collections.abc.Callable[[], None]]: ...
